+++ README.md
@@ -1,3 +1,3 @@
 # recipes_api
 
- MONGO_URI="mongodb://user:password@localhost:27017/test?authSource=admin" MONGO_DATABASE=demo go run main.go
+MONGO_URI="mongodb://user:password@localhost:27017/test?authSource=admin" MONGO_DATABASE=demo go run main.go
